Conflicts:
	scripts/alsa-lib

+++ helper_functions.sh
@@ -62,10 +62,11 @@ download() {
         echo "Download cache dir ${DOWNLOAD_CACHE_DIR} doesn't exist - not using it" 2>&1
     fi
     if [ -n "${DOWNLOAD_CACHE_DIR}" -a \
-         -f "${DOWNLOAD_CACHE_DIR}/${FILENAME}" ] ; then
+         -f "${DOWNLOAD_CACHE_DIR}/${FILENAME}" -a \
+         ! -f "${FILENAME}" ] ; then
         echo "File ${FILENAME} exists in ${DOWNLOAD_CACHE_DIR} - copying from there"
         cp "${DOWNLOAD_CACHE_DIR}/${FILENAME}" "${FILENAME}"
-    elif [ -f $FILENAME ] ; then
+    elif [ -f "$FILENAME" ] ; then
         echo "Skipping download of ${FILENAME} - already present"
     else
         wget --no-check-certificate "$SOURCE" -O ${FILENAME}
@@ -251,14 +252,6 @@ do_build_install() {
     do_configure
     do_make
     do_install "$1"
-}
-
-# Builds a package, and puts all of its installation into
-# a single archive
-do_simple() {
-    do_build_install "$1"
-
-    build_package ${NAME} ${VERSION} "$1"
 }
 
 # Build using CMake
